requirements.txt
openai>=1.12.0
langgraph>=0.0.32
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
typing-extensions>=4.5.0
pydantic>=2.0.0
pandas 